Fix merge and import errors
