+++ optimum/exporters/openvino/model_configs.py
@@ -405,65 +405,6 @@ class Qwen3OpenVINOConfig(TextDecoderWithPositionIdsOpenVINOConfig):
             common_inputs = super().inputs
         return common_inputs
 
-class DummyQwen3VLLMInputGenerator(DummyTextInputGenerator):
-    SUPPORTED_INPUT_NAMES = (
-        "input_ids",
-        "attention_mask",
-        "token_type_ids",
-        "position_ids",
-        "visual_pos_masks",
-        "deepstack_visual_embeds",
-    )
-
-    def __init__(
-        self,
-        task: str,
-        normalized_config: NormalizedTextConfig,
-        batch_size: int = DEFAULT_DUMMY_SHAPES["batch_size"],
-        sequence_length: int = DEFAULT_DUMMY_SHAPES["sequence_length"],
-        num_choices: int = DEFAULT_DUMMY_SHAPES["num_choices"],
-        random_batch_size_range: Optional[Tuple[int, int]] = None,
-        random_sequence_length_range: Optional[Tuple[int, int]] = None,
-        random_num_choices_range: Optional[Tuple[int, int]] = None,
-        padding_side: str = "right",
-        **kwargs,
-    ):
-        super().__init__(
-            task=task,
-            normalized_config=normalized_config,
-            batch_size=batch_size,
-            sequence_length=sequence_length,
-            num_choices=num_choices,
-            random_batch_size_range=random_batch_size_range,
-            random_sequence_length_range=random_sequence_length_range,
-            random_num_choices_range=random_num_choices_range,
-            padding_side=padding_side,
-            **kwargs,
-        )
-        self.embed_dim = normalized_config.hidden_size
-        self.num_layers = len(self.normalized_config.deepstack_visual_indexes)
-
-    def generate(
-        self,
-        input_name: str,
-        framework: str = "pt",
-        int_dtype: str = "int64",
-        float_dtype: str = "fp32",
-        bool_dtype: str = "bool",
-    ):
-        if input_name == "deepstack_visual_embeds":
-            return self.random_float_tensor(
-                [self.num_layers, 2 * self.sequence_length, self.embed_dim], framework=framework, dtype=float_dtype
-            )
-        if input_name == "visual_pos_masks":
-            return self.constant_tensor(
-                shape=[self.batch_size, self.sequence_length],
-                framework=framework,
-                value=1,
-                dtype=DTYPE_MAPPER.pt(bool_dtype),
-            )
-        return super().generate(input_name, framework, int_dtype, float_dtype)
-
 
 @register_in_tasks_manager(
     "qwen3_vl_text",
@@ -2203,99 +2144,6 @@ class LlavaQwen2OpenVINOConfig(BaseVLMOpenVINOConfig):
         return super().rename_ambiguous_inputs(inputs)
 
 
-class PooledProjectionsDummyInputGenerator(DummyInputGenerator):
-    SUPPORTED_INPUT_NAMES = ["pooled_projections"]
-
-    def __init__(
-        self,
-        task: str,
-        normalized_config: NormalizedConfig,
-        batch_size: int = DEFAULT_DUMMY_SHAPES["batch_size"],
-        random_batch_size_range: Optional[Tuple[int, int]] = None,
-        **kwargs,
-    ):
-        self.task = task
-        self.batch_size = batch_size
-        config = normalized_config.config
-        pooled_projection_dim = getattr(config, "pooled_projection_dim", None)
-        if pooled_projection_dim is None and hasattr(config, "get"):
-            pooled_projection_dim = config.get("pooled_projection_dim", None)
-
-        self.pooled_projection_dim = pooled_projection_dim
-
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if self.pooled_projection_dim is None:
-            raise ValueError("`pooled_projection_dim` is required to generate `pooled_projections` input.")
-        shape = [self.batch_size, self.pooled_projection_dim]
-        return self.random_float_tensor(shape, framework=framework, dtype=float_dtype)
-
-
-class DummyTransformerTimestpsInputGenerator(DummyTimestepInputGenerator):
-    SUPPORTED_INPUT_NAMES = ("timestep", "text_embeds", "time_ids", "timestep_cond", "guidance")
-
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if input_name in ["timestep", "guidance"]:
-            shape = [self.batch_size]
-            return self.random_float_tensor(shape, max_value=self.vocab_size, framework=framework, dtype=float_dtype)
-        return super().generate(input_name, framework, int_dtype, float_dtype)
-
-
-class DummyUnetVisionInputGenerator(DummyVisionInputGenerator):
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if input_name not in ["sample", "latent_sample"]:
-            return super().generate(input_name, framework, int_dtype, float_dtype)
-        # add height and width discount for enable any resolution generation
-        return self.random_float_tensor(
-            shape=[self.batch_size, self.num_channels, self.height - 1, self.width - 1],
-            framework=framework,
-            dtype=float_dtype,
-        )
-
-
-class DummyUnetTimestepInputGenerator(DummyTimestepInputGenerator):
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if input_name != "timestep":
-            return super().generate(input_name, framework, int_dtype, float_dtype)
-        shape = [self.batch_size]
-        return self.random_int_tensor(shape, max_value=self.vocab_size, framework=framework, dtype=int_dtype)
-
-
-class DummySanaTimestepInputGenerator(DummyTimestepInputGenerator):
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if input_name != "timestep":
-            return super().generate(input_name, framework, int_dtype, float_dtype)
-        shape = [self.batch_size]
-        return self.random_int_tensor(shape, max_value=self.vocab_size, framework=framework, dtype=float_dtype)
-
-
-class DummyUnetEncoderInputGenerator(DummySeq2SeqDecoderTextInputGenerator):
-    def __init__(
-        self,
-        task: str,
-        normalized_config: NormalizedTextConfig,
-        batch_size: int = DEFAULT_DUMMY_SHAPES["batch_size"],
-        sequence_length: int = DEFAULT_DUMMY_SHAPES["sequence_length"],
-        num_choices: int = DEFAULT_DUMMY_SHAPES["num_choices"],
-        random_batch_size_range: Optional[Tuple[int, int]] = None,
-        random_sequence_length_range: Optional[Tuple[int, int]] = None,
-        random_num_choices_range: Optional[Tuple[int, int]] = None,
-        **kwargs,
-    ):
-        super().__init__(
-            task,
-            normalized_config,
-            batch_size=batch_size,
-            sequence_length=sequence_length,
-            num_choices=num_choices,
-            random_batch_size_range=random_batch_size_range,
-            random_sequence_length_range=random_sequence_length_range,
-            random_num_choices_range=random_num_choices_range,
-            **kwargs,
-        )
-        if hasattr(normalized_config.config, "model_max_length"):
-            self.sequence_length = normalized_config.config.model_max_length
-
-
 @register_in_tasks_manager("unet", *["semantic-segmentation"], library_name="diffusers")
 @register_in_tasks_manager("unet-2d-condition", *["semantic-segmentation"], library_name="diffusers")
 class UNetOpenVINOConfig(VisionOpenVINOConfig):
@@ -2561,118 +2409,6 @@ class DcaeDecoderOpenVINOConfig(VisionOpenVINOConfig):
         return {
             "sample": {0: "batch_size", 2: "height", 3: "width"},
         }
-
-
-def _get_flux_ids_dim(config) -> int:
-    for attr_name in ("axes_dims_rope", "axes_dim", "axes_dims"):
-        value = getattr(config, attr_name, None)
-        if value is not None:
-            if isinstance(value, (list, tuple)):
-                return len(value)
-            if isinstance(value, int):
-                return value
-
-    if hasattr(config, "get"):
-        for key in ("axes_dims_rope", "axes_dim", "axes_dims"):
-            value = config.get(key, None)
-            if value is not None:
-                if isinstance(value, (list, tuple)):
-                    return len(value)
-                if isinstance(value, int):
-                    return value
-
-    return 3
-
-
-class DummyFluxTransformerInputGenerator(DummyVisionInputGenerator):
-    SUPPORTED_INPUT_NAMES = (
-        "pixel_values",
-        "pixel_mask",
-        "sample",
-        "latent_sample",
-        "hidden_states",
-        "img_ids",
-    )
-
-    def __init__(
-        self,
-        task: str,
-        normalized_config: NormalizedVisionConfig,
-        batch_size: int = DEFAULT_DUMMY_SHAPES["batch_size"],
-        num_channels: int = DEFAULT_DUMMY_SHAPES["num_channels"],
-        width: int = DEFAULT_DUMMY_SHAPES["width"] // 4,
-        height: int = DEFAULT_DUMMY_SHAPES["height"] // 4,
-        **kwargs,
-    ):
-        super().__init__(task, normalized_config, batch_size, num_channels, width, height, **kwargs)
-        if getattr(normalized_config, "in_channels", None):
-            self.num_channels = normalized_config.in_channels // 4
-        self.ids_dim = _get_flux_ids_dim(normalized_config.config)
-
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if input_name in ["hidden_states", "sample"]:
-            shape = [self.batch_size, (self.height // 2) * (self.width // 2), self.num_channels * 4]
-            return self.random_float_tensor(shape, framework=framework, dtype=float_dtype)
-        if input_name == "img_ids":
-            img_ids_height = self.height // 2
-            img_ids_width = self.width // 2
-            return self.random_int_tensor(
-                (
-                    [self.batch_size, img_ids_height * img_ids_width, self.ids_dim]
-                    if is_diffusers_version("<", "0.31.0")
-                    else [img_ids_height * img_ids_width, self.ids_dim]
-                ),
-                min_value=0,
-                max_value=min(img_ids_height, img_ids_width),
-                framework=framework,
-                dtype=float_dtype,
-            )
-
-        return super().generate(input_name, framework, int_dtype, float_dtype)
-
-
-class DummyFluxTextInputGenerator(DummySeq2SeqDecoderTextInputGenerator):
-    SUPPORTED_INPUT_NAMES = (
-        "decoder_input_ids",
-        "decoder_attention_mask",
-        "encoder_outputs",
-        "encoder_hidden_states",
-        "txt_ids",
-    )
-
-    def __init__(
-        self,
-        task: str,
-        normalized_config: NormalizedTextConfig,
-        batch_size: int = DEFAULT_DUMMY_SHAPES["batch_size"],
-        sequence_length: int = DEFAULT_DUMMY_SHAPES["sequence_length"],
-        random_batch_size_range: Optional[Tuple[int, int]] = None,
-        random_sequence_length_range: Optional[Tuple[int, int]] = None,
-        **kwargs,
-    ):
-        super().__init__(
-            task=task,
-            normalized_config=normalized_config,
-            batch_size=batch_size,
-            sequence_length=sequence_length,
-            random_batch_size_range=random_batch_size_range,
-            random_sequence_length_range=random_sequence_length_range,
-            **kwargs,
-        )
-        self.ids_dim = _get_flux_ids_dim(normalized_config.config)
-
-    def generate(self, input_name: str, framework: str = "pt", int_dtype: str = "int64", float_dtype: str = "fp32"):
-        if input_name == "txt_ids":
-            import torch
-
-            shape = (
-                [self.batch_size, self.sequence_length, self.ids_dim]
-                if is_diffusers_version("<", "0.31.0")
-                else [self.sequence_length, self.ids_dim]
-            )
-            dtype = DTYPE_MAPPER.pt(float_dtype)
-            return torch.full(shape, 0, dtype=dtype)
-        return super().generate(input_name, framework, int_dtype, float_dtype)
 
 
 @register_in_tasks_manager("flux-transformer", *["semantic-segmentation"], library_name="diffusers")
